API's argument types are changed to uintptr_t from uint64_t

+++ include/collection.h
@@ -5,10 +5,10 @@
 
 #define GD_COLLECTION_API \
 	RetCode_t	(*isEmpty)(void* self); \
-	RetCode_t	(*contains)(void* self, uint64_t element); \
-	RetCode_t	(*add)(void* self, uint64_t element); \
-	uint64_t	(*get)(void* self, uint32_t index); \
-	RetCode_t	(*remove)(void* self, uint64_t element); \
+	RetCode_t	(*contains)(void* self, uintptr_t element); \
+	RetCode_t	(*add)(void* self, uintptr_t element); \
+	uintptr_t	(*get)(void* self, uint32_t index); \
+	RetCode_t	(*remove)(void* self, uintptr_t element); \
 	RetCode_t	(*clear)(void* self);
 
 #define GD_COLLECTION_PROPS \
