+++ include/queue.h
@@ -5,8 +5,8 @@
 
 struct GDQueue {
 	GD_COLLECTION_PROPS
-	RetCode_t	(*push)(void* self, uint64_t element);
-	uint64_t	(*pop)(void* self);
+	RetCode_t	(*push)(void* self, uintptr_t element);
+	uintptr_t	(*pop)(void* self);
 	void		(*dump)(void* self);
 	int32_t		head;
 	int32_t		tail;
